feat: improve compose fp style and refactors folders
